requirements.txt
torch==2.0.1
transformers==4.30.2
keybert==0.7.0
pillow==10.0.0
pytesseract==0.3.10
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
pyinstaller==5.13.0
tkinter==0.1.0